First source commit
I was developing this privately in the WebService::Google::Client namespace but gave up trying to collaborate with the holder of the cpan package space. Retaining it privately seemed like a waste and was in some cases problematic so I settled on the new namespace and pushed to cpan. I will clean up over the coming days and continue to work on refining.