CPANEL RELEASE V0.21
Includes refactored api_query including the ulimate inclusion of parameters specified with location='query' to be added as GET params. Performed an audit of existing endpoints to cofirm that can always assume ok to append the '?'. The other hackable options remain available but default behavious is now working thanks to @rabbiveesh